Do not reference Grynwald.MdDocs assembly from DemoProject
It seems like there is a compiler bug that was triggered by upgrading the Microsoft.Extensions.Configuration packages in Grynwald.MdDocs.Common.
Because the upgrade brings in a new version of System.Text.Json which defines System.ObsoleteAttribute as internal class, the compiler can no longer resolve the reference to the attribute in the XML documentation comments (e.g. <see cref=".." />).
This seems to be occurrence of the bug described here: https://github.com/dotnet/roslyn/issues/4033

Removing the reference from the demo project (we need to keep the project reference so projects are built in the right order) fixes the problem, since System.Text.Json is no longer included in the references of that project.